Batched rendering (BlockPos)
Split the Outline class into a few different classes depending on functionality. Abstracts away some functionality for different implementations of handling vertices. If an outline is part of block space (i.e. cannot move, and it's transform is locked to the world grid) then it is a batched outline.
Added a way of rendering the main batched vertex buffer which will be used for block space outlines. Rendering this is very fast, but updating the buffer needs a lot of work as it currently uses a rather crude implementation.