Optimise outline preperation
Outline preperation is now significantly faster. Outlines to remove, suspend and regenerate because of merge changes are now all done in the same loop, and the merge logic itself is now much more efficient as when a merge outline is added or removed, all of its overlapping outlines are marked as dirty, meaning we now just regenerate all dirty outlines, without needing to search for which ones are actually affected.