Line mesh builder
Added a function which builds cuboid lines into the mesh builder, this is now used to create nice boxes around entities. This is combined with the new inflation parameter in outlines which specify how much extra padding around the object the outline should have. Added a utility method for getting a colour from a hex code.