Greedy outline mesher
Modified the outline mesh builder class to have the option to allow greedy meshing to be used, this drastically reduces the number of vertices in a generated outline, especially one which contains lots of long lines, but does come with an upfront generation time cost. This will probably be worked around in future with the ability to generate outlines asynchronously, but this should only really be used for outlines which are very large, which should be a rarity.