Outline merging
Added the ability for outlines to visually merge if they share a merge key, and merging is enabled. The outline which was present first will have the edges of the part where outlines join. Outline merging is still quite slow, so implementing an async system for outline generation will be worth it. Also fixed a bug where outline regeneration order was inconsistent due to sets being unordered. It will now follow the order in which they were queued.