Merged Outline + Collision Rework
This is still work in progress, merged outlines aren't quite right yet, but collision outlines work perfectly. Building the block group mesh now uses a new and simpler method, we now create 2 different clusters, one for collisions and one for the actual mesh. For each cluster we pass in the relevant block set which we calculate beforehand. This does cause some overlap of shared edges, but it shouldn't be visible, just a little bit wasteful, this could potentially be optimised in the future.
Fixed a bug where outlines could be queued to be regenerated multiple times in the same frame, causing unnecessary computation.
Fixed a bug where if an outline had no vertices due to overlapping a mesh with show collisions turned off, it would be regenerated every frame, along with any colliding outlines.